fixed sonar issues
