+++ src/resources/reference-data/ReferenceDataShow.tsx
@@ -12,7 +12,8 @@ import { ValueField } from '../projects/ProjectShow'
 import { R_MEDIA_TYPE, type ResourceTypes } from '../../constants'
 import useCanAccess from '../../hooks/useCanAccess'
 import HistoryButton from '../../components/HistoryButton'
-import { getResourceName } from '.'
+import { getResourceName } from './utils'
+import { useConfigData } from '../../utils/useConfigData'
 
 interface ShowActionProps {
   resource: ResourceTypes
@@ -45,7 +46,8 @@ export default function ReferenceDataShow({
   name
 }: PropType): React.ReactElement {
   const cName: ResourceTypes = name as ResourceTypes
-  const resourceName = getResourceName(name)
+  const configData = useConfigData()
+  const resourceName = getResourceName(name, configData)
   return (
     <Show resource={name} actions={<Actions resource={cName} />}>
       <Typography variant='h5' fontWeight='bold' sx={{ padding: '15px' }}>
